refactor: upgrade code in eload_submission, preparation code in eload_preparation
