+++ eva_submission/submission_config.py
@@ -6,7 +6,6 @@ from ebi_eva_common_pyutils.config import Configuration, cfg
 from ebi_eva_common_pyutils.logger import logging_config as log_cfg
 
 from eva_submission import __version__
-from eva_submission.xlsx.xlsx_parser_eva import EvaXlsxReader
 
 logger = log_cfg.get_logger(__name__)
 
@@ -22,71 +21,6 @@ class EloadConfig(Configuration):
             # in that case the first search path is set to be the config files
             self.config_file = search_path[0]
             pass
-
-    def upgrade_if_needed(self, analysis_alias):
-        """
-        Upgrades unversioned configs (i.e. pre-1.0) to the current version, making a backup first and using the
-        provided analysis alias for all vcf files. Currently doesn't perform any other version upgrades.
-        """
-        if 'version' not in self.content:
-            logger.info(f'No version found in config, upgrading to version {__version__}.')
-            self.backup()
-
-            self.set('version', value=__version__)
-            if 'submission' not in self.content:
-                logger.error('Need submission config section to upgrade')
-                logger.error('Try running prepare_submission or prepare_backlog_study to build a config from scratch.')
-                raise ValueError('Need submission config section to upgrade')
-
-            # Note: if we're converting an old config, there's only one analysis
-            if not analysis_alias:
-                analysis_alias = self._get_analysis_alias_from_metadata()
-            analysis_data = {
-                'assembly_accession': self.pop('submission', 'assembly_accession'),
-                'assembly_fasta': self.pop('submission', 'assembly_fasta'),
-                'assembly_report': self.pop('submission', 'assembly_report'),
-                'vcf_files': self.pop('submission', 'vcf_files')
-            }
-            analysis_dict = {analysis_alias: analysis_data}
-            self.set('submission', 'analyses', value=analysis_dict)
-
-            if 'validation' in self.content:
-                self.pop('validation', 'valid', 'vcf_files')
-                self.set('validation', 'valid', 'analyses', value=analysis_dict)
-
-            if 'brokering' in self.content:
-                brokering_vcfs = {
-                    vcf_file: index_dict
-                    for vcf_file, index_dict in self.pop('brokering', 'vcf_files').items()
-                }
-                analysis_dict[analysis_alias]['vcf_files'] = brokering_vcfs
-                self.set('brokering', 'analyses', value=analysis_dict)
-                analysis_accession = self.pop('brokering', 'ena', 'ANALYSIS')
-                self.set('brokering', 'ena', 'ANALYSIS', analysis_alias, value=analysis_accession)
-
-        else:
-            # TODO think through how complicated this might get...
-            logger.info(f"Config is version {self.query('version')}, not upgrading.")
-
-    def _get_analysis_alias_from_metadata(self):
-        """
-        Returns analysis alias only if we find a metadata spreadsheet and it has exactly one analysis.
-        Otherwise provides an error message and raise an error.
-        """
-        metadata_spreadsheet = self.query('validation', 'valid', 'metadata_spreadsheet')
-        if metadata_spreadsheet:
-            reader = EvaXlsxReader(metadata_spreadsheet)
-            if len(reader.analysis) == 1:
-                return reader.analysis[0].get('Analysis Alias')
-
-            if len(reader.analysis) > 1:
-                logger.error("Can't assign analysis alias: multiple analyses found in metadata!")
-            else:
-                logger.error("Can't assign analysis alias: no analyses found in metadata!")
-        else:
-            logger.error("Can't assign analysis alias: no metadata found!")
-        logger.error("Try running upgrade_config and passing an analysis alias explicitly.")
-        raise ValueError("Can't find an analysis alias for config upgrade.")
 
     def backup(self):
         if self.config_file and self.content and os.path.isdir(os.path.dirname(self.config_file)):
@@ -120,6 +54,9 @@ class EloadConfig(Configuration):
 
     def clear(self):
         self.content = {}
+
+    def __contains__(self, item):
+        return item in self.content
 
     def __setitem__(self, item, value):
         """Allow dict-style write access, e.g. config['this']='that'."""
