+++ tests/test_eload_submission.py
@@ -1,105 +1,45 @@
-import glob
 import os
-import shutil
+from copy import deepcopy
 from unittest import TestCase
-
-from ebi_eva_common_pyutils.config import cfg
+from unittest.mock import patch
 
 from eva_submission import ROOT_DIR
-from eva_submission.eload_submission import EloadPreparation
-from eva_submission.submission_config import load_config
-from eva_submission.xlsx.xlsx_parser_eva import EvaXlsxReader, EvaXlsxWriter
-
-
-def touch(filepath, content=None):
-    with open(filepath, 'w') as open_file:
-        if content:
-            open_file.write(content)
+from eva_submission.eload_submission import Eload
+from eva_submission.submission_config import EloadConfig, load_config
 
 
 class TestEload(TestCase):
     resources_folder = os.path.join(ROOT_DIR, 'tests', 'resources')
 
-    def setUp(self) -> None:
+    def setUp(self):
         config_file = os.path.join(self.resources_folder, 'submission_config.yml')
         load_config(config_file)
         # Need to set the directory so that the relative path set in the config file works from the top directory
         os.chdir(ROOT_DIR)
-        self.eload = EloadPreparation(1)
+        self.eload = Eload(55)
+        self.updated_config = EloadConfig(os.path.join(self.eload.eload_dir, 'updated_config.yml'))
+        # Used to restore test config after each test
+        self.original_cfg = deepcopy(self.eload.eload_cfg.content)
 
-    def tearDown(self) -> None:
-        eloads = glob.glob(os.path.join(self.resources_folder, 'eloads', 'ELOAD_1'))
-        for eload in eloads:
-            shutil.rmtree(eload)
-        genomes = glob.glob(os.path.join(self.resources_folder, 'genomes'))
-        for genome in genomes:
-            shutil.rmtree(genome)
+    def tearDown(self):
+        self.eload.eload_cfg.content = self.original_cfg
+        if os.path.exists(f'{self.eload.eload_cfg.config_file}.old'):
+            os.remove(f'{self.eload.eload_cfg.config_file}.old')
 
-    def create_vcfs(self, num_files=2):
-        paths = []
-        for i in range(num_files):
-            vcf = os.path.join(self.eload.eload_dir, '10_submitted', 'vcf_files', f'file{i}.vcf')
-            touch(vcf)
-            paths.append(vcf)
-        return paths
+    def test_upgrade_config(self):
+        """Tests config upgrade for a post-brokering config."""
+        self.eload.upgrade_config_if_needed('analysis alias')
+        self.assertEqual(self.updated_config.content, self.eload.eload_cfg.content)
 
-    def create_metadata(self, num_analyses=0):
-        source_metadata = os.path.join(self.resources_folder, 'metadata.xlsx')
-        metadata = os.path.join(self.eload.eload_dir, '10_submitted', 'metadata_file', 'metadata.xlsx')
-        shutil.copyfile(source_metadata, metadata)
-        if num_analyses:
-            writer = EvaXlsxWriter(metadata)
-            writer.set_analysis([
-                {
-                    header: f'analysis{i}'
-                    for header in
-                    ['Analysis Title', 'Analysis Alias', 'Description', 'Project Title', 'Experiment Type', 'Reference']
-                } for i in range(num_analyses)
-            ])
-            writer.save()
-        return metadata
+    def test_upgrade_config_no_analysis_alias_passed(self):
+        """If no analysis alias is passed, it should retrieve from metadata when possible."""
+        with patch('eva_submission.eload_submission.EvaXlsxReader') as mock_reader:
+            mock_reader.return_value.analysis = [{'Analysis Alias': 'analysis alias'}]
+            self.eload.upgrade_config_if_needed(analysis_alias=None)
+            self.assertEqual(self.updated_config.content, self.eload.eload_cfg.content)
 
-    def test_copy_from_ftp(self):
-        assert os.listdir(os.path.join(self.eload.eload_dir, '10_submitted', 'vcf_files')) == []
-        assert os.listdir(os.path.join(self.eload.eload_dir, '10_submitted', 'metadata_file')) == []
-        self.eload.copy_from_ftp(1, 'john')
-        assert os.listdir(os.path.join(self.eload.eload_dir, '10_submitted', 'vcf_files')) == ['data.vcf.gz']
-        assert os.listdir(os.path.join(self.eload.eload_dir, '10_submitted', 'metadata_file')) == ['metadata.xlsx']
-
-    def test_detect_submitted_metadata(self):
-        self.create_vcfs()
-        metadata = self.create_metadata()
-
-        self.eload.detect_submitted_metadata()
-        self.eload.check_submitted_filenames()
-        # Check that the metadata spreadsheet is in the config file
-        assert self.eload.eload_cfg.query('submission', 'metadata_spreadsheet') == metadata
-
-    def test_check_submitted_filenames_multiple_analyses(self):
-        # create some extra vcf files and analyses
-        vcfs = self.create_vcfs(num_files=5)
-        self.create_metadata(num_analyses=2)
-
-        self.eload.detect_submitted_metadata()
-        with self.assertRaises(ValueError):
-            self.eload.check_submitted_filenames()
-
-    def test_replace_values_in_metadata(self):
-        metadata = self.create_metadata()
-
-        reader = EvaXlsxReader(metadata)
-        assert reader.project['Tax ID'] == 9606
-        assert reader.analysis[0]['Reference'] == 'GCA_000001405.1'
-        self.eload.replace_values_in_metadata(taxid=10000, reference_accession='GCA_000009999.9')
-        reader = EvaXlsxReader(metadata)
-        assert reader.project['Tax ID'] == 10000
-        assert reader.analysis[0]['Reference'] == 'GCA_000009999.9'
-
-    def test_find_genome_single_sequence(self):
-        cfg.content['eutils_api_key'] = None
-        self.eload.eload_cfg.set('submission', 'scientific_name', value='Thingy thingus')
-        self.eload.eload_cfg.set('submission', 'analyses', 'Analysis alias test', 'assembly_accession', value='AJ312413.2')
-        self.eload.find_genome()
-        assert self.eload.eload_cfg['submission']['analyses']['Analysis alias test']['assembly_fasta'] \
-               == 'tests/resources/genomes/thingy_thingus/AJ312413.2/AJ312413.2.fa'
-        assert 'assembly_report' not in self.eload.eload_cfg['submission']
+    def test_upgrade_config_already_updated(self):
+        """An already up-to-date config shouldn't get modified."""
+        original_content = deepcopy(self.updated_config.content)
+        self.eload.upgrade_config_if_needed('analysis alias')
+        self.assertEqual(original_content, self.updated_config.content)
